+++ src/main/java/gdsc/be/mount/storage/service/FolderService.java
@@ -2,10 +2,7 @@ package gdsc.be.mount.storage.service;
 
 import gdsc.be.mount.storage.Enum.FileFolderType;
 import gdsc.be.mount.storage.dto.request.FileFolderUpdateRequest;
-import gdsc.be.mount.storage.dto.request.FileUploadRequest;
 import gdsc.be.mount.storage.dto.request.FolderCreateRequest;
-import gdsc.be.mount.storage.dto.response.FileDownloadResponse;
-import gdsc.be.mount.storage.dto.response.FileUploadResponse;
 import gdsc.be.mount.storage.dto.response.FolderCreateResponse;
 import gdsc.be.mount.storage.dto.response.FolderInfoResponse;
 import gdsc.be.mount.storage.entity.FileFolder;
@@ -17,13 +14,8 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.UrlResource;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.util.UriUtils;
 
 import java.io.IOException;
-import java.net.MalformedURLException;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.List;
 import java.util.UUID;
@@ -32,114 +24,12 @@ import java.util.UUID;
 @RequiredArgsConstructor
 @Transactional
 @Slf4j
-public class FileFolderService {
+public class FolderService {
 
     private final FileFolderRepository fileFolderRepository;
 
     @Value("${upload.path}")
     private String uploadPath;
-
-    public FileUploadResponse uploadFile(MultipartFile file, FileUploadRequest fileUploadRequest) {
-
-        validate(file); // 파일 유효성 검사
-
-        Long parentId = fileUploadRequest.parentId();
-        String userName = fileUploadRequest.userName();
-
-        // 만약 부모의 폴더의 주인이 자신이 아니라면 예외 발생
-        checkIfParentIsYours(parentId, userName);
-
-        // 만약 parentId 가 폴더가 아니라면 예외 발생
-        checkIfParentIsFolder(parentId);
-
-        String originalFileName = file.getOriginalFilename(); // 사용자가 등록한 최초 파일명
-        String storeFileName = createStoreFileName(originalFileName); // 서버 내부에서 관리할 파일명
-        String logicalFilePath = getFullLogicalPath(userName, storeFileName, parentId); // 파일의 논리적 경로
-        String physicalFilePath = uploadPath + storeFileName; // 파일의 물리적 경로
-
-        log.debug("[uploadFile] originalFileName: {}, logicalFilePath: {}", originalFileName, logicalFilePath);
-
-        try {
-            // 1. 파일 시스템에서 물리적 파일 저장
-            savePhysicalFile(file, physicalFilePath);
-
-            try {
-                // 2. DB 에 파일 메타데이터 저장
-                FileFolder savedFileFolder = saveFileMetadataToDB(fileUploadRequest, originalFileName, storeFileName, logicalFilePath, file.getSize(), file.getContentType());
-
-                // 3. 부모 폴더에 자식 폴더 id 추가
-                if(parentId != null){
-                    addChildIdIntoParentFolder(parentId, savedFileFolder.getId());
-                }
-
-                return FileUploadResponse.fromEntity(savedFileFolder);
-            } catch (Exception dbException) {
-                // 만약 DB에 파일 메타데이터 저장 중에 예외가 발생하면 물리적 파일 삭제 후 예외 다시 던지기
-                deletePhysicalFile(physicalFilePath);
-                throw dbException;
-            }
-        } catch (IOException ex) {
-            throw FileUploadException.EXCEPTION;
-        }
-    }
-
-    public Long deleteFile(Long fileId, String userName) {
-        // 파일 확인 및 권한 검사
-        FileFolder fileFolder = getFileFolderForDeletionAfterCheck(fileId, userName);
-
-        // 삭제하려는 대상이 파일인지 확인
-        if(fileFolder.getFileFolderType() == FileFolderType.FOLDER){
-            throw FileDeleteNotAllowedException.EXCEPTION;
-        }
-
-        String deletedFileName = fileFolder.getOriginalName();
-        String physicalFilePath = uploadPath + fileFolder.getStoredName(); // 가상 폴더 구조이므로 가상 경로가 아닌 물리적인 실제 경로를 사용
-        log.debug("[deleteFile] FileName: {}", deletedFileName);
-
-        try {
-            // 1. DB 에서 파일 메타데이터 삭제
-            deleteFileMetadata(fileId);
-
-            // 2. 파일 시스템에서 물리적 파일 삭제
-            deletePhysicalFile(physicalFilePath);
-
-            return fileFolder.getId();
-        } catch (IOException ex) {
-            throw FileDeletionException.EXCEPTION;
-        }
-    }
-
-    public FileDownloadResponse downloadFile(Long fileId, String userName) {
-        // 파일 확인 및 권한 검사
-        FileFolder fileFolder = getFileFolderForDownloadAfterCheck(fileId, userName);
-
-        // 다운로드 하려는 대상이 파일인지 확인
-        if(fileFolder.getFileFolderType() == FileFolderType.FOLDER){
-            throw FileDownloadNotAllowedException.EXCEPTION;
-        }
-
-        String originalFileName = fileFolder.getOriginalName();
-        String saveFileName = fileFolder.getStoredName();
-        String physicalFilePath = uploadPath + saveFileName; // 가상 폴더 구조이므로 가상 경로가 아닌 물리적인 실제 경로를 사용
-
-        try {
-
-            UrlResource resource = getResource(physicalFilePath);
-
-            log.debug("[downloadFile] saveFileName: {}, URL Resource: {}", saveFileName, resource);
-
-            // 다운로드 시 가독성 위해 최초 파일명 사용
-            String encodedOriginalFileName = UriUtils.encode(originalFileName, StandardCharsets.UTF_8);
-            String contentDisposition = "attachment; filename=\"" + encodedOriginalFileName + "\"";
-
-            return new FileDownloadResponse(resource, contentDisposition);
-        } catch (MalformedURLException ex) {
-            // URL 생성 오류
-            throw FileDownloadExpcetion.EXCEPTION;
-        } catch (IOException ex) {
-            throw FileDownloadExpcetion.EXCEPTION;
-        }
-    }
 
     public FolderCreateResponse createFolder(FolderCreateRequest folderCreateRequest) {
 
@@ -222,39 +112,6 @@ public class FileFolderService {
 
     // ====================================================================================================
 
-    /*
-    파일 저장 관련 메서드
-     */
-
-    private void validate(MultipartFile file) {
-        // 파일이 존재하는지 확인
-        if (file == null) {
-            throw FileUploadException.EXCEPTION;
-        }
-
-        // 파일명이 비어있는지 확인
-        if (file.getOriginalFilename() == null || file.getOriginalFilename().isEmpty()) {
-            throw FileUploadException.EXCEPTION;
-        }
-
-        // 파일 크기가 0인지 확인
-        if (file.getSize() == 0) {
-            throw FileUploadException.EXCEPTION;
-        }
-    }
-
-    private String createStoreFileName(String originalFileName){
-        // 원본 파일명에서 확장자 추출
-        String ext = extractExt(originalFileName);
-
-        // 확장자가 없는 경우 기본 확장자를 사용 (예. txt 로 설정)
-        if (ext.isEmpty()) {
-            ext = "txt";
-        }
-
-        return UUID.randomUUID().toString().substring(0, 5) + "." + ext;
-    }
-
     private String extractExt(String originalFilename) {
         // 확장자 별도 추출
         int pos = originalFilename.lastIndexOf(".");
@@ -292,36 +149,11 @@ public class FileFolderService {
         return parentFileFolder.getPath();
     }
 
-    private void savePhysicalFile(MultipartFile file, String filePath) throws IOException {
-        file.transferTo(Files.createFile(Path.of(filePath)));
-    }
-
-    private FileFolder saveFileMetadataToDB(FileUploadRequest fileUploadRequest, String originalFileName, String storeFileName, String logicalFilePath, long fileSize, String fileType) {
-        return fileFolderRepository.save(fileUploadRequest.toEntity(originalFileName, storeFileName, logicalFilePath, fileSize, fileType));
-    }
-
     private void addChildIdIntoParentFolder(Long parentId, Long childId) {
         FileFolder parentFileFolder = fileFolderRepository.findById(parentId).orElseThrow();
         parentFileFolder.addChildId(childId);
         fileFolderRepository.save(parentFileFolder);
     }
-
-    /*
-    파일 삭제 관련 메서드
-     */
-
-    private void deleteFileMetadata(Long fileId) {
-        fileFolderRepository.deleteById(fileId);
-    }
-
-    private void deletePhysicalFile(String filePath) throws IOException {
-        Path fileToDelete = Path.of(filePath);
-        Files.delete(fileToDelete);
-    }
-
-    /*
-    파일 다운로드 관련 메서드
-     */
 
     UrlResource getResource(String path) throws IOException {
         Path filePath = Path.of(path);
@@ -335,10 +167,6 @@ public class FileFolderService {
         return resource;
     }
 
-    /*
-    폴더 생성 관련 메서드
-     */
-
     private static String generateRandomFolderName() {
         // 랜덤한 UUID를 사용하여 폴더 이름 생성
         return UUID.randomUUID().toString().substring(0, 5);
@@ -348,8 +176,8 @@ public class FileFolderService {
         return fileFolderRepository.save(folderCreateRequest.toEntity(folderName, folderDir, parentId, userName));
     }
 
-    /*
-     파일 및 폴더 권한 검사 관련 메서드
+    /**
+     * 파일 및 폴더 권한 검사 관련 메서드
      */
     private FileFolder getFileFolderFromDatabase(Long fileId) {
         return fileFolderRepository.findById(fileId)
@@ -386,18 +214,6 @@ public class FileFolderService {
                 throw FileUploadException.EXCEPTION;
             }
         }
-    }
-
-    private FileFolder getFileFolderForDownloadAfterCheck(Long fileId, String userName) {
-        FileFolder fileFolder = getFileFolderFromDatabase(fileId);
-        checkOwnership(userName, fileFolder.getUserName(),false, true, false);
-        return fileFolder;
-    }
-
-    private FileFolder getFileFolderForDeletionAfterCheck(Long fileId, String userName) {
-        FileFolder fileFolder = getFileFolderFromDatabase(fileId);
-        checkOwnership(userName, fileFolder.getUserName(),false, false, false);
-        return fileFolder;
     }
 
     private FileFolder getFileFolderForUpdateAfterCheck(Long fileId, String userName) {
